Refactor stats class inside geospatial
