+++ src/main/java/org/opensearch/geospatial/action/upload/geojson/Uploader.java
@@ -17,8 +17,8 @@ import org.opensearch.action.bulk.BulkItemResponse;
 import org.opensearch.action.bulk.BulkRequestBuilder;
 import org.opensearch.action.bulk.BulkResponse;
 import org.opensearch.common.collect.MapBuilder;
-import org.opensearch.geospatial.action.upload.UploadMetric;
-import org.opensearch.geospatial.action.upload.UploadStats;
+import org.opensearch.geospatial.stats.upload.UploadMetric;
+import org.opensearch.geospatial.stats.upload.UploadStats;
 
 /**
  * Uploader will upload GeoJSON objects from UploadGeoJSONRequestContent as
