+++ src/test/java/org/opensearch/geospatial/GeospatialTestHelper.java
@@ -37,9 +37,9 @@ import org.opensearch.action.support.replication.ReplicationResponse;
 import org.opensearch.common.Randomness;
 import org.opensearch.common.UUIDs;
 import org.opensearch.common.collect.Tuple;
-import org.opensearch.geospatial.action.upload.UploadMetric;
 import org.opensearch.geospatial.action.upload.geojson.ContentBuilder;
 import org.opensearch.geospatial.action.upload.geojson.UploadGeoJSONRequestContent;
+import org.opensearch.geospatial.stats.upload.UploadMetric;
 import org.opensearch.index.shard.ShardId;
 import org.opensearch.test.OpenSearchTestCase;
 import org.opensearch.test.RandomObjects;
