+++ src/test/java/org/opensearch/geospatial/stats/UploadMetricTests.java
@@ -3,13 +3,14 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package org.opensearch.geospatial.action.upload;
+package org.opensearch.geospatial.stats;
 
 import static org.opensearch.geospatial.GeospatialTestHelper.buildFieldNameValuePair;
 import static org.opensearch.geospatial.GeospatialTestHelper.randomLowerCaseString;
 
 import org.opensearch.common.Strings;
 import org.opensearch.geospatial.GeospatialTestHelper;
+import org.opensearch.geospatial.stats.upload.UploadMetric;
 import org.opensearch.test.OpenSearchTestCase;
 
 public class UploadMetricTests extends OpenSearchTestCase {
